fixing build brakes and small refactoring
