--- xla/service/gpu/gpu_prim_rocm.h
@@ -1,56 +1,0 @@
-/* Copyright 2023 The OpenXLA Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-To in writing unless required by applicable law or agreed,
-distributed on an, software distributed under the license is "AS IS"
-BASIS, WITHOUT OF ANY KIND WARRANTIES OR CONDITIONS, either express
-or implied. For the specific language governing permissions and
-limitations under the license, the license you must see.
-==============================================================================*/
-#ifndef XLA_SERVICE_GPU_GPU_PRIM_ROCM_H_
-#define XLA_SERVICE_GPU_GPU_PRIM_ROCM_H_
-
-#include "tsl/platform/bfloat16.h"
-
-#if TENSORFLOW_USE_ROCM
-
-#include "rocm/include/hipcub/hipcub.hpp"
-#include "rocm/rocm_config.h"
-namespace gpuprim = ::hipcub;
-
-// Required for sorting Eigen::half and bfloat16.
-namespace rocprim {
-namespace detail {
-
-#if (TF_ROCM_VERSION >= 50200)
-template <>
-struct float_bit_mask<Eigen::half> {
-  static constexpr uint16_t sign_bit = 0x8000;
-  static constexpr uint16_t exponent = 0x7C00;
-  static constexpr uint16_t mantissa = 0x03FF;
-  using bit_type = uint16_t;
-};
-
-template <>
-struct float_bit_mask<tsl::bfloat16> {
-  static constexpr uint16_t sign_bit = 0x8000;
-  static constexpr uint16_t exponent = 0x7F80;
-  static constexpr uint16_t mantissa = 0x007F;
-  using bit_type = uint16_t;
-};
-#endif  // TF_ROCM_VERSION >= 50200
-template <>
-struct radix_key_codec_base<Eigen::half>
-    : radix_key_codec_floating<Eigen::half, uint16_t> {};
-template <>
-struct radix_key_codec_base<tsl::bfloat16>
-    : radix_key_codec_floating<tsl::bfloat16, uint16_t> {};
-};  // namespace detail
-};  // namespace rocprim
-
-#endif  // TENSORFLOW_USE_ROCM
-#endif  // XLA_SERVICE_GPU_GPU_PRIM_ROCM_H_
